+++ common/src/main/java/traben/entity_texture_features/features/property_reading/properties/generic_properties/StringArrayOrRegexProperty.java
@@ -15,6 +15,8 @@ public abstract class StringArrayOrRegexProperty extends RandomProperty {
     protected final ObjectOpenHashSet<String> ARRAY;
     protected final RegexAndPatternPropertyMatcher MATCHER;
 
+    protected final boolean usesRegex;
+
 
     protected StringArrayOrRegexProperty(String string) throws RandomPropertyException {
         ORIGINAL_INPUT = string;
@@ -24,6 +26,7 @@ public abstract class StringArrayOrRegexProperty extends RandomProperty {
                 || string.startsWith("iregex:") || string.startsWith("ipattern:")) {
             MATCHER = getStringMatcher_Regex_Pattern_List_Single(string);
             ARRAY = ObjectOpenHashSet.of(string);
+            usesRegex = true;
         } else {
             String[] array = string.trim().split("\\s+");
 
@@ -35,8 +38,26 @@ public abstract class StringArrayOrRegexProperty extends RandomProperty {
                     array) {
                 ARRAY.add(shouldForceLowerCaseCheck() ? str.toLowerCase() : str);
             }
-            MATCHER = ARRAY::contains;
+            MATCHER = this::testArray;
+            usesRegex = false;
         }
+    }
+
+    private boolean testArray(String fromEntity){
+        boolean matches = false;
+        for (String string : ARRAY) {
+            if (string.startsWith("!")) {
+                matches = true;
+                if (string.substring(1).equals(fromEntity)) {
+                    matches = false;
+                    break;
+                }
+            } else if (string.equals(fromEntity)) {
+                matches = true;
+                break;
+            }
+        }
+        return matches;
     }
 
     @Nullable